requirements.txt
streamlit
pytesseract
python-docx
PyPDF2
ollama
langgraph
requests
Pillow
spacy
faiss-cpu
scikit-learn
langchain
python-dotenv
langchain-openai
langchain-core
langchain-community
pdf2image
langchain_groq
